Transferred genetics stuff to a new package.
